Adding FMPI2C to F4xx family.
This adds the FMPI2C i2c devices but it was a bit difficult due to the
fact that only one i2c version is allowed. This was fixed by adding a
symlink to i2c_v2.yaml which was named i2c2_v2.yaml this allows for
both devices to exist on the same chip and keeps the yaml files aligned.