Move pin cleanup from code to YAML, before adding more clean-up rules
This commit does not introduce any PAC changes, the results are identical
except that now all pins are sorted, also the ones read from data/extras.
I wanted to do this in an extra commit to make reviewing easier, although
it doesn't add value at the moment and is just a first step.