Override pin names for OPAMPs
Some peripherals, such as OPAMPs, have PIN names and variants that
are distributed across the chip XMLs and IP XMLs. It would be very
cumbersome to reconstruct these completely by combining this information.
If there are only a few versions of these IPs, it may make more sense
to define them manually in YAML files.

This commit adds this mechanism and implements it consistently for
all OPAMPs.